refactor(upload): Stabilise Github Error
